Fix unpack/clean script for freshly cloned repo

+++ tools/cleanBalatroFolder.mjs
@@ -5,13 +5,15 @@ import chalk from 'chalk';
 
 const srcFolder = './balatroSrc';
 
-fs.readdirSync(srcFolder).forEach((file) => {
-  if (fs.lstatSync(path.join(srcFolder, file)).isDirectory()) {
-    fs.rmdirSync(path.join(srcFolder, file), { recursive: true });
-    return;
-  }
+if (fs.existsSync(srcFolder)) {
+  fs.readdirSync(srcFolder).forEach((file) => {
+    if (fs.lstatSync(path.join(srcFolder, file)).isDirectory()) {
+      fs.rmdirSync(path.join(srcFolder, file), { recursive: true });
+      return;
+    }
 
-  fs.rmSync(path.join(srcFolder, file));
-});
+    fs.rmSync(path.join(srcFolder, file));
+  });
+}
 
 console.log(chalk.green(`-> \`.balatroSrc/\` folder cleaned.`));
